Make req.route an object saving route and baseUrl
When using an error handler you have access to req.route
which describes the most recently matched route. However,
some other informaiton (like req.baseUrl) gets changed
before the error handler runs.

This commit changes req.route to an object that also contains
the req.baseUrl before it gets changed.

Since the actual route object is not scoped to the current
event, this commit instead uses a new object to store the
baseUrl but sets the __proto__ to the route object so that
all its properties are also available.